% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTP.R
\name{ARTP}
\alias{ARTP}
\title{Main function, based on Yu et al.
Initial programming: Malte Braitmaier}
\usage{
ARTP(daten, J, n_perm)
}
\arguments{
\item{daten}{A matrix, rows contain variables,
columns contain the original p-value followed}

\item{J}{Truncation point}

\item{n_perm}{Number of permutations}
}
\value{
A data frame p.values.groups with the p value for each
        group and truncation
}
\description{
Main function, based on Yu et al.
Initial programming: Malte Braitmaier
}
