% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get.p.value.cov}
\alias{get.p.value.cov}
\title{Get p-value for each covariate}
\usage{
get.p.value.cov(x, y, adjust_vars)
}
\description{
Applies the ARTP to binary data given the observed covariates and the
response. The number of observations is n. The number of covariates
is m.
}
\seealso{
artp.fit
}
