% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get.p.value}
\alias{get.p.value}
\title{Get p-value for each covariate considering original and permuted outcome}
\usage{
get.p.value(
  X,
  y,
  n.permutations,
  n.cov,
  cov_ind,
  adjust_vars,
  parallel = FALSE,
  nc = 3,
  verbose = FALSE
)
}
\description{
Applies the ARTP to binary data given the observed covariates and the
response. The number of observations is n. The number of covariates
is m.
}
\seealso{
artp.fit
}
