% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.per.group}
\alias{predict.per.group}
\title{Predict outcome per group of covariates}
\usage{
\method{predict}{per.group}(fit, X.new, alpha, predict.all)
}
\arguments{
\item{fit}{The output of fit.artp}

\item{X.new}{The new observation matrix (n x m)}

\item{alpha}{The significance level, no default}

\item{predict.all}{To predict outcome based on all groups (Default = FALSE)}
}
\description{
Predict outcome per group of covariates
}
\seealso{
artp.predict
}
