% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{artp.fit}
\alias{artp.fit}
\title{Apply the ARTP to Binary Data}
\usage{
artp.fit(X, y, groups, trunc.point = 5, n.permutations = 50)
}
\arguments{
\item{X}{Binary matrix of size n x m}

\item{y}{Binaryy response vector of length n}

\item{groups}{List of groups. Each item is a vector with the
indices of the covariates that belong to that group}

\item{trunc.point}{The truncation point used (Default = 5)}

\item{n.permutations}{Number of permutations (Default = 50)}
}
\value{
A data frame p.values.groups with the p value for each
        group
}
\description{
Applies the ARTP to binary data given the observed covariates and the
response. The number of observations is n. The number of covariates
is m.
}
