% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{artp.fit}
\alias{artp.fit}
\title{Apply the ARTP to Binary Data}
\usage{
artp.fit(
  X,
  y,
  groups,
  adjust_vars = NULL,
  trunc.point = 5,
  n.permutations = 50,
  verbose = FALSE,
  single_covariates = TRUE,
  parallel = FALSE,
  nc = 3
)
}
\arguments{
\item{X}{Binary matrix of size n x m, adjustment variables are could be continous (e.g., age)}

\item{y}{Binary response vector of length n}

\item{groups}{List of groups. Each item is a vector with the
indices of the covariates that belong to that group}

\item{adjust_vars}{Indices of adjustment variables for the regression (Default = NULL)}

\item{trunc.point}{The truncation point used (Default = 5)}

\item{n.permutations}{Number of permutations (Default = 50)}

\item{verbose}{If TRUE, shows progress bar (Default = FALSE),
FALSE in case of parallel <- TRUE}

\item{single_covariates}{If TRUE, covariates that do not belong to a group, get
their own individual groups (Default = TRUE)}

\item{parallel}{Boolean, whether to use parallel:mcapply (Default = FALSE)}

\item{nc}{Number of cores to use for parallel:mcmapply (Default = 3)}
}
\value{
A list of input parameters or artp.predict, including a
        data frame p.values.groups with the p value for each group
}
\description{
Applies the ARTP to binary data given the observed covariates and the
response. The number of observations is n. The number of covariates
is m.
}
