% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{artp.predict}
\alias{artp.predict}
\title{Predict the Outcome with the ARTP}
\usage{
artp.predict(fit, X.new, alpha = 0.1)
}
\arguments{
\item{fit}{The output of fit.artp}

\item{X.new}{The new observation matrix}

\item{alpha}{The significance level}
}
\value{
The predictions for the outcome y
}
\description{
Predict the Outcome with the ARTP
}
